+++ manual/source/version_history.rst
@@ -158,6 +158,8 @@ General Updates
   Geant4 names (e.g. "e-").
 * Print out extent of loaded world when using an external geometry file.
 * **EMD** physics has a minimum applicable kinetic energy of 1 MeV to prevent crashes in Geant4.
+* Optional executable argument added to ptc2bdsim to control ROOT split-level of sampler branches. Same
+  functionality as the BDSIM option `samplersSplitLevel`.
 
 Bug Fixes
 ---------